Refactor frontend QueueApi to MatchApi
