+++ frontend/src/components/MatchMaking/MatchingModal.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { v4 as uuidv4 } from 'uuid';
-import { exitQueue } from '../../api/QueueApi.js';
+import { exitQueue } from '../../api/MatchApi.js';
 import { getCookie } from '../../utils/helpers.js';
 import Queue from './Queue';
 import '../../css/Modal.css'
